pull code

+++ src/main/java/com/tale/service/ContentsService.java
@@ -1,23 +1,21 @@
 package com.tale.service;
 
+import com.blade.exception.ValidatorException;
 import com.blade.ioc.annotation.Bean;
 import com.blade.ioc.annotation.Inject;
 import com.blade.kit.DateKit;
 import com.blade.kit.StringKit;
-import com.tale.exception.TipException;
-import com.tale.init.TaleConst;
 import com.tale.model.dto.Types;
 import com.tale.model.entity.Comments;
 import com.tale.model.entity.Contents;
 import com.tale.model.entity.Relationships;
-import com.tale.utils.TaleUtils;
+import com.tale.model.params.ArticleParam;
 import com.vdurmont.emoji.EmojiParser;
 import io.github.biezhi.anima.Anima;
+import io.github.biezhi.anima.core.AnimaQuery;
 import io.github.biezhi.anima.page.Page;
-import io.github.biezhi.anima.page.PageRow;
 
-import java.util.Optional;
-
+import static com.tale.bootstrap.TaleConst.SQL_QUERY_ARTICLES;
 import static io.github.biezhi.anima.Anima.deleteById;
 import static io.github.biezhi.anima.Anima.select;
 
@@ -38,15 +36,19 @@ public class ContentsService {
      *
      * @param id 唯一标识
      */
-    public Optional<Contents> getContents(String id) {
+    public Contents getContents(String id) {
+        Contents contents = null;
         if (StringKit.isNotBlank(id)) {
             if (StringKit.isNumber(id)) {
-                return Optional.ofNullable(select().from(Contents.class).byId(id));
+                contents = select().from(Contents.class).byId(id);
             } else {
-                return Optional.ofNullable(select().from(Contents.class).where(Contents::getSlug, id).one());
+                contents = select().from(Contents.class).where(Contents::getSlug, id).one();
+            }
+            if (null != contents) {
+                return this.mapContent(contents);
             }
         }
-        return Optional.empty();
+        return contents;
     }
 
     /**
@@ -55,40 +57,8 @@ public class ContentsService {
      * @param contents 文章对象
      */
     public Integer publish(Contents contents) {
-        if (null == contents) {
-            throw new TipException("文章对象为空");
-        }
-        if (StringKit.isBlank(contents.getTitle())) {
-            throw new TipException("文章标题不能为空");
-        }
-        if (contents.getTitle().length() > TaleConst.MAX_TITLE_COUNT) {
-            throw new TipException("文章标题最多可以输入" + TaleConst.MAX_TITLE_COUNT + "个字符");
-        }
-
-        if (StringKit.isBlank(contents.getContent())) {
-            throw new TipException("文章内容不能为空");
-        }
-        // 最多可以输入5w个字
-        int len = contents.getContent().length();
-        if (len > TaleConst.MAX_TEXT_COUNT) {
-            throw new TipException("文章内容最多可以输入" + TaleConst.MAX_TEXT_COUNT + "个字符");
-        }
         if (null == contents.getAuthorId()) {
-            throw new TipException("请登录后发布文章");
-        }
-
-        if (StringKit.isNotBlank(contents.getSlug())) {
-            if (contents.getSlug().length() < 5) {
-                throw new TipException("路径太短了");
-            }
-            if (!TaleUtils.isPath(contents.getSlug())) {
-                throw new TipException("您输入的路径不合法");
-            }
-
-            long count = new Contents().where("type", contents.getType()).and("slug", contents.getSlug()).count();
-            if (count > 0) {
-                throw new TipException("该路径已经存在，请重新输入");
-            }
+            throw new ValidatorException("请登录后发布文章");
         }
 
         contents.setContent(EmojiParser.parseToAliases(contents.getContent()));
@@ -96,6 +66,7 @@ public class ContentsService {
         int time = DateKit.nowUnix();
         contents.setCreated(time);
         contents.setModified(time);
+        contents.setHits(0);
 
         String tags       = contents.getTags();
         String categories = contents.getCategories();
@@ -120,11 +91,12 @@ public class ContentsService {
         contents.setTags(contents.getTags() != null ? contents.getTags() : "");
         contents.setCategories(contents.getCategories() != null ? contents.getCategories() : "");
 
-        String  tags       = contents.getTags();
-        String  categories = contents.getCategories();
-        Integer cid        = contents.getCid();
+        Integer cid = contents.getCid();
 
         contents.updateById(cid);
+
+        String tags       = contents.getTags();
+        String categories = contents.getCategories();
 
         if (null != contents.getType() && !contents.getType().equals(Types.PAGE)) {
             Anima.delete().from(Relationships.class).where(Relationships::getCid, cid).execute();
@@ -140,12 +112,12 @@ public class ContentsService {
      * @param cid 文章id
      */
     public void delete(int cid) {
-        Optional<Contents> contents = this.getContents(cid + "");
-        contents.ifPresent(content -> {
+        Contents contents = this.getContents(cid + "");
+        if (null != contents) {
             deleteById(Contents.class, cid);
             Anima.delete().from(Relationships.class).where(Relationships::getCid, cid).execute();
             Anima.delete().from(Comments.class).where(Comments::getCid, cid).execute();
-        });
+        }
     }
 
     /**
@@ -157,10 +129,40 @@ public class ContentsService {
      * @return
      */
     public Page<Contents> getArticles(Integer mid, int page, int limit) {
-        String sql = "select a.* from t_contents a left join t_relationships b on a.cid = b.cid " +
-                "where b.mid = "+mid+" and a.status = 'publish' and a.type = 'post' order by a.created desc";
-        
-        return select().from(Contents.class).page(sql, page, limit);
+        return select().bySQL(Contents.class, SQL_QUERY_ARTICLES, mid).page(page, limit);
     }
 
+    public Page<Contents> findArticles(ArticleParam articleParam) {
+        AnimaQuery<Contents> query = select().from(Contents.class).exclude(Contents::getContent);
+
+        if (StringKit.isNotEmpty(articleParam.getStatus())) {
+            query.and(Contents::getStatus, articleParam.getStatus());
+        }
+
+        if (StringKit.isNotEmpty(articleParam.getTitle())) {
+            query.and(Contents::getTitle).like("%" + articleParam.getTitle() + "%");
+        }
+
+        if (StringKit.isNotEmpty(articleParam.getCategories())) {
+            query.and(Contents::getCategories).like("%" + articleParam.getCategories() + "%");
+        }
+
+        query.and(Contents::getType, articleParam.getType());
+        query.order(articleParam.getOrderBy());
+        Page<Contents> articles = query.page(articleParam.getPage(), articleParam.getLimit());
+        return articles.map(this::mapContent);
+    }
+
+    private Contents mapContent(Contents contents) {
+        if (StringKit.isNotEmpty(contents.getSlug())) {
+            String url = "/" + contents.getSlug();
+            contents.setUrl(url.replaceAll("[/]+", "/"));
+        } else {
+            contents.setUrl("/article/" + contents.getCid());
+        }
+        if (StringKit.isNotEmpty(contents.getContent())) {
+            contents.setContent(contents.getContent().replaceAll("\\\\\"", "\\\""));
+        }
+        return contents;
+    }
 }
